headlessscms/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
contentful==1.13.1
Django==4.0.4
django-ckeditor-5==0.0.15
django-cors-headers==3.11.0
django-editorjs-fields==0.2.6
django-filter==21.1
django-mapbox-location-field==2.0.0
django-parler==2.3
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.1.3
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.4
martor==1.6.13
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2021.3
pyuploadcare==3.0.0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.9
zipp==3.8.0
